requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
firebase-admin==6.5.0
cryptography==42.0.8
Pillow==10.3.0

# NumPy por versión de Python
numpy==1.26.4; python_version >= "3.9" and python_version < "3.13"
numpy==1.24.4; python_version < "3.9"

# Elige UNO de estos (CPU o GPU):
onnxruntime==1.18.0            # CPU
# onnxruntime-gpu==1.18.0      # GPU (CUDA 12+)

# Opcional (solo si vas a extraer embeddings reales aquí)
# insightface==0.7.3
